More v6 work

+++ shared/services/rewards/types.go
@@ -42,7 +42,7 @@ type MinipoolInfo struct {
 	EndSlot                 uint64
 	AttestationScore        *big.Int
 	CompletedAttestations   map[uint64]bool
-	AttestationCount        uint64
+	AttestationCount        int
 }
 
 type IntervalDutiesInfo struct {
